move PackageName to separate module
